third_party/WebKit/LayoutTests/platform/win/fast/spatial-navigation/snav-div-scrollable-but-without-focusable-content-expected.txt





PASS successfullyParsed is true

TEST COMPLETE
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
FAIL gFocusedDocument.activeElement.getAttribute("id") should be end. Was start.
FAIL gFocusedDocument.activeElement.getAttribute("id") should be end. Was start.
FAIL gFocusedDocument.activeElement.getAttribute("id") should be end. Was start.
FAIL gFocusedDocument.activeElement.getAttribute("id") should be end. Was start.
FAIL gFocusedDocument.activeElement.getAttribute("id") should be end. Was start.
FAIL gFocusedDocument.activeElement.getAttribute("id") should be end. Was start.
FAIL gFocusedDocument.activeElement.getAttribute("id") should be end. Was start.
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
This test is to test that a scrollable div can scroll and reveal its content even if it does not have any focusable content.
